% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{head.varpred}
\alias{head.varpred}
\title{Return the First or Last Part of a varpred object}
\usage{
\method{head}{varpred}(x, ...)
}
\arguments{
\item{x}{varpred object.}

\item{...}{additional aurguments passed to head.}
}
\description{
Return the First or Last Part of a varpred object
}
