% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{tail.varpred}
\alias{tail.varpred}
\title{Return the First or Last Part of a varpred object}
\usage{
\method{tail}{varpred}(x, ...)
}
\arguments{
\item{x}{varpred object.}

\item{...}{additional aurguments passed to tail.}
}
\description{
Return the First or Last Part of a varpred object
}
