% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsfuns.R
\name{varpredtheme}
\alias{varpredtheme}
\title{Customized theme for varpred plots}
\usage{
varpredtheme(type = c("minimal", "complete"))
}
\description{
Sets a theme for varpred and other ggplot objects
}
\examples{
library(ggplot2)
varpredtheme()
set.seed(4567)
x <- rnorm(100, 3, 5)
y <- 0.4 + 0.7*x + rnorm(100)
df <- data.frame(y = y, x = x)
m1 <- lm(y ~ x, df)

pred1 <- varpred(m1, "x")
plot(pred1)

}
